refactor: 패키지명 vo -> model로 변경
