+++ Api/src/main/java/com/dingdong/api/community/service/CommunityService.java
@@ -15,8 +15,8 @@ import com.dingdong.api.idcard.dto.KeywordDto;
 import com.dingdong.core.exception.BaseException;
 import com.dingdong.domain.common.util.SliceUtil;
 import com.dingdong.domain.domains.community.adaptor.CommunityAdaptor;
-import com.dingdong.domain.domains.community.domain.Community;
-import com.dingdong.domain.domains.community.domain.CommunityImage;
+import com.dingdong.domain.domains.community.domain.entity.Community;
+import com.dingdong.domain.domains.community.domain.model.CommunityImage;
 import com.dingdong.domain.domains.community.domain.strategy.GenerateCommunityInvitationCodeStrategy;
 import com.dingdong.domain.domains.community.validator.CommunityValidator;
 import com.dingdong.domain.domains.idcard.adaptor.IdCardAdaptor;
