+++ Api/src/main/java/com/dingdong/api/idcard/dto/UserInfoDto.java
@@ -1,7 +1,7 @@
 package com.dingdong.api.idcard.dto;
 
 
-import com.dingdong.domain.domains.idcard.domain.vo.UserInfo;
+import com.dingdong.domain.domains.idcard.domain.model.UserInfo;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 import lombok.Getter;
