+++ Api/src/main/java/com/dingdong/api/idcard/service/IdCardService.java
@@ -8,13 +8,13 @@ import com.dingdong.api.idcard.dto.CreateKeywordDto;
 import com.dingdong.api.idcard.dto.IdCardDetailsDto;
 import com.dingdong.api.idcard.dto.KeywordDto;
 import com.dingdong.domain.domains.community.adaptor.CommunityAdaptor;
-import com.dingdong.domain.domains.community.domain.Community;
+import com.dingdong.domain.domains.community.domain.entity.Community;
 import com.dingdong.domain.domains.idcard.adaptor.IdCardAdaptor;
 import com.dingdong.domain.domains.idcard.domain.entity.IdCard;
 import com.dingdong.domain.domains.idcard.domain.entity.Keyword;
 import com.dingdong.domain.domains.idcard.validator.IdCardValidator;
 import com.dingdong.domain.domains.image.adaptor.ImageAdaptor;
-import com.dingdong.domain.domains.image.domain.DeleteImage;
+import com.dingdong.domain.domains.image.domain.entity.DeleteImage;
 import com.dingdong.domain.domains.user.domain.User;
 import java.util.List;
 import java.util.Objects;
