+++ Domain/src/main/java/com/dingdong/domain/domains/community/repository/AdminRepository.java
@@ -1,7 +1,7 @@
 package com.dingdong.domain.domains.community.repository;
 
 
-import com.dingdong.domain.domains.community.domain.Admin;
+import com.dingdong.domain.domains.community.domain.entity.Admin;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AdminRepository extends JpaRepository<Admin, Long> {}
