+++ Domain/src/main/java/com/dingdong/domain/domains/community/repository/CommunityRepository.java
@@ -1,7 +1,7 @@
 package com.dingdong.domain.domains.community.repository;
 
 
-import com.dingdong.domain.domains.community.domain.Community;
+import com.dingdong.domain.domains.community.domain.entity.Community;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
