+++ Domain/src/main/java/com/dingdong/domain/domains/idcard/domain/entity/IdCard.java
@@ -3,8 +3,8 @@ package com.dingdong.domain.domains.idcard.domain.entity;
 
 import com.dingdong.domain.domains.AbstractTimeStamp;
 import com.dingdong.domain.domains.idcard.domain.enums.CharacterType;
-import com.dingdong.domain.domains.idcard.domain.vo.Character;
-import com.dingdong.domain.domains.idcard.domain.vo.UserInfo;
+import com.dingdong.domain.domains.idcard.domain.model.Character;
+import com.dingdong.domain.domains.idcard.domain.model.UserInfo;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
