+++ Domain/src/main/java/com/dingdong/domain/domains/idcard/domain/model/UserInfo.java
@@ -1,4 +1,4 @@
-package com.dingdong.domain.domains.idcard.domain.vo;
+package com.dingdong.domain.domains.idcard.domain.model;
 
 
 import javax.persistence.Embeddable;
