+++ Domain/src/main/java/com/dingdong/domain/domains/image/adaptor/ImageAdaptor.java
@@ -2,7 +2,7 @@ package com.dingdong.domain.domains.image.adaptor;
 
 
 import com.dingdong.core.annotation.Adaptor;
-import com.dingdong.domain.domains.image.domain.DeleteImage;
+import com.dingdong.domain.domains.image.domain.entity.DeleteImage;
 import com.dingdong.domain.domains.image.repository.ImageRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
