+++ Domain/src/main/java/com/dingdong/domain/domains/image/repository/ImageRepository.java
@@ -1,7 +1,7 @@
 package com.dingdong.domain.domains.image.repository;
 
 
-import com.dingdong.domain.domains.image.domain.DeleteImage;
+import com.dingdong.domain.domains.image.domain.entity.DeleteImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
