+++ Domain/src/main/java/com/dingdong/domain/domains/user/domain/User.java
@@ -2,8 +2,8 @@ package com.dingdong.domain.domains.user.domain;
 
 
 import com.dingdong.domain.domains.AbstractTimeStamp;
-import com.dingdong.domain.domains.community.domain.Community;
-import com.dingdong.domain.domains.idcard.domain.vo.Character;
+import com.dingdong.domain.domains.community.domain.entity.Community;
+import com.dingdong.domain.domains.idcard.domain.model.Character;
 import com.dingdong.domain.domains.user.domain.enums.GenderType;
 import java.util.ArrayList;
 import java.util.List;
